refactor(FTT): rename file to conform to snake case
